:truck: move Slack views to the views file
