+++ src/models/platform/message.ts
@@ -2,4 +2,5 @@ export type Message = string | View | InteractiveView
 
 export abstract class View {}
 
+// TODO: add interface methods (gratitudeMessage, coffeeRoulette...)
 export abstract class InteractiveView {}
